Virtual-Voice-Assistant/requirements.txt
pyttsx3
SpeechRecognition
tensorflow
Keras
Keras_Preprocessing
numpy
scikit-learn
Pillow
stability-sdk
newsapi-python
wolframalpha
wikipedia
speedtest-cli
youtube-search-python
appopener
WMI
pynput
requests
psutil
python-dotenv
PyAudio
pycaw
winshell
pywin32
